requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
anyio==4.2.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
decorator==5.1.1
etelemetry==0.3.1
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.1
genai==2.1.0
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
looseversion==1.3.0
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.5
networkx==3.2.1
nibabel==5.2.0
nipype==1.8.6
numpy==1.26.4
openai==0.27.10
packaging==23.2
pandas==2.2.0
parso==0.8.3
pathlib==1.0.1
pdf2image==1.17.0
pexpect==4.9.0
pillow==10.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
prov==2.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydeck==0.8.1b0
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.37.0
streamlit==1.31.0
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.3.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
traits==6.3.2
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
uvicorn==0.27.0.post1
validators==0.22.0
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
